+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ * Basilisp modules now resolve lazily as needed; decoupling them from pythons module hierarchy. (#957)
+
 ### Added
  * Added several missing functions to `basilisp.core` (#956)